+++ specutils/io/default_loaders/tabular_fits.py
@@ -1,10 +1,10 @@
 import logging
 import os
+import _io
 
 import numpy as np
 
 from astropy.io import fits
-from astropy.nddata import StdDevUncertainty
 from astropy.table import Table
 import astropy.units as u
 from astropy.wcs import WCS
@@ -18,25 +18,42 @@ __all__ = ['tabular_fits_loader', 'tabular_fits_writer']
 
 
 def identify_tabular_fits(origin, *args, **kwargs):
-    # check if file can be opened with this reader
     # args[0] = filename
-    with fits.open(args[0]) as hdulist:
-        # Test if fits has extension of type BinTable and check against
-        # known keys of already defined specific formats
-        return (len(hdulist) > 1 and
-                isinstance(hdulist[1], fits.BinTableHDU) and not
-                (fits.getheader(args[0]).get('TELESCOP') == 'MULTI' and
-                 fits.getheader(args[0]).get('HLSPACRN') == 'MUSCLES' and
-                 fits.getheader(args[0]).get('PROPOSID') == 13650) and not
-                (fits.getheader(args[0]).get('TELESCOP') == 'SDSS 2.5-M' and
-                 fits.getheader(args[0]).get('FIBERID') > 0) and not
-                (fits.getheader(args[0]).get('TELESCOP') == 'HST' and
-                 fits.getheader(args[0]).get('INSTRUME') in ('COS', 'STIS')) and not
-                 fits.getheader(args[0]).get('TELESCOP') == 'JWST')
+    hdu = kwargs.get('hdu', 1)
+    # Check if filename conforms to naming convention and writer has not been
+    # asked to write to primary (IMAGE-only) HDU
+    if origin == 'write':
+        return (hdu > 0 and args[0].endswith(('.fits', '.fit')) and not
+                args[0].endswith(('wcs.fits', 'wcs1d.fits', 'wcs.fit')))
+
+    if isinstance(args[2], fits.hdu.hdulist.HDUList):
+        hdulist = args[2]
+    elif isinstance(args[2], _io.BufferedReader):
+        hdulist = fits.open(args[2])
+    else:  # if isinstance(args[2], _io.FileIO):
+        hdulist = fits.open(args[0], **kwargs)
+
+    # Test if fits has extension of type BinTable and check against
+    # known keys of already defined specific formats
+    is_tab = (len(hdulist) > 1 and
+              isinstance(hdulist[hdu], fits.BinTableHDU) and not
+              (hdulist[0].header.get('TELESCOP') == 'MULTI' and
+               hdulist[0].header.get('HLSPACRN') == 'MUSCLES' and
+               hdulist[0].header.get('PROPOSID') == 13650) and not
+              (hdulist[0].header.get('TELESCOP') == 'SDSS 2.5-M' and
+               hdulist[0].header.get('FIBERID') > 0) and not
+              (hdulist[0].header.get('TELESCOP') == 'HST' and
+               hdulist[0].header.get('INSTRUME') in ('COS', 'STIS')) and not
+              hdulist[0].header.get('TELESCOP') == 'JWST')
+
+    if not isinstance(args[2], (fits.hdu.hdulist.HDUList, _io.BufferedReader)):
+        hdulist.close()
+
+    return is_tab
 
 
 @data_loader("tabular-fits", identifier=identify_tabular_fits,
-             dtype=Spectrum1D, extensions=['fits'])
+             dtype=Spectrum1D, extensions=['fits', 'fit'], priority=6)
 def tabular_fits_loader(file_obj, column_mapping=None, hdu=1, **kwargs):
     """
     Load spectrum from a FITS file.
@@ -89,7 +106,7 @@ def tabular_fits_loader(file_obj, column_mapping=None, hdu=1, **kwargs):
 
 
 @custom_writer("tabular-fits")
-def tabular_fits_writer(spectrum, file_name, update_header=False, **kwargs):
+def tabular_fits_writer(spectrum, file_name, hdu=1, update_header=False, **kwargs):
     """
     Write spectrum to BINTABLE extension of a FITS file.
 
@@ -98,11 +115,22 @@ def tabular_fits_writer(spectrum, file_name, update_header=False, **kwargs):
     spectrum: Spectrum1D
     file_name: str
         The path to the FITS file
+    hdu: int
+        Header Data Unit in FITS file to write to (currently only extension HDU 1)
     update_header: bool
         Update FITS header with all compatible entries in `spectrum.meta`
+    wunit : str or `~astropy.units.Unit`
+        Unit for the spectral axis (wavelength or frequency-like)
+    funit : str or `~astropy.units.Unit`
+        Unit for the flux (and associated uncertainty)
+    wtype : str or `~numpy.dtype`
+        Floating point type for storing spectral axis array
+    ftype : str or `~numpy.dtype`
+        Floating point type for storing flux array
     """
-    flux = spectrum.flux
-    disp = spectrum.spectral_axis
+    if hdu < 1:
+        raise ValueError(f'FITS does not support BINTABLE extension in HDU {hdu}.')
+
     header = spectrum.meta.get('header', fits.header.Header()).copy()
 
     if update_header:
@@ -115,16 +143,28 @@ def tabular_fits_writer(spectrum, file_name, update_header=False, **kwargs):
     for keyword in ['NAXIS', 'NAXIS1', 'NAXIS2']:
         header.remove(keyword, ignore_missing=True)
 
+    # Add dispersion array and unit
+    wtype = kwargs.pop('wtype', spectrum.spectral_axis.dtype)
+    wunit = u.Unit(kwargs.pop('wunit', spectrum.spectral_axis.unit))
+    disp = spectrum.spectral_axis.to(wunit, equivalencies=u.spectral())
+
     # Mapping of spectral_axis types to header TTYPE1
-    dispname = disp.unit.physical_type
+    dispname = wunit.physical_type
     if dispname == "length":
         dispname = "wavelength"
 
-    columns = [disp, flux]
+    # Add flux array and unit
+    ftype = kwargs.pop('ftype', spectrum.flux.dtype)
+    funit = u.Unit(kwargs.pop('funit', spectrum.flux.unit))
+    flux = spectrum.flux.to(funit, equivalencies=u.spectral_density(disp))
+
+    columns = [disp.astype(wtype), flux.astype(ftype)]
     colnames = [dispname, "flux"]
+
     # Include uncertainty - units to be inferred from spectrum.flux
     if spectrum.uncertainty is not None:
-        columns.append(spectrum.uncertainty.quantity)
+        unc = spectrum.uncertainty.quantity.to(funit, equivalencies=u.spectral_density(disp))
+        columns.append(unc.astype(ftype))
         colnames.append("uncertainty")
 
     # For 2D data transpose from row-major format
@@ -142,4 +182,6 @@ def tabular_fits_writer(spectrum, file_name, update_header=False, **kwargs):
 
     tab = Table(columns, names=colnames, meta=header)
 
+    # Todo: support writing to other HDUs than the default (1st)
+    #       and an 'update' mode so different HDUs can be written to separately
     tab.write(file_name, format="fits", **kwargs)